erc721

+++ src/pages/primitives/index.html.ts
@@ -25,7 +25,7 @@ const html = `<p>Here we introduce you to some primitive data types available in
         uint256 ranges from 0 to 2 ** 256 - 1
     */</span>
     <span class="hljs-keyword">uint8</span> <span class="hljs-keyword">public</span> u8 = <span class="hljs-number">1</span>;
-    <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> u256 = <span class="hljs-number">456</span>;
+    <span class="hljs-keyword">uint256</span> <span class="hljs-keyword">public</span> u256 = <span class="hljs-number">456</span>;
     <span class="hljs-keyword">uint</span> <span class="hljs-keyword">public</span> u = <span class="hljs-number">123</span>; <span class="hljs-comment">// uint is an alias for uint256</span>
 
     <span class="hljs-comment">/*
@@ -36,7 +36,7 @@ const html = `<p>Here we introduce you to some primitive data types available in
     int128 ranges from -2 ** 127 to 2 ** 127 - 1
     */</span>
     <span class="hljs-keyword">int8</span> <span class="hljs-keyword">public</span> i8 = <span class="hljs-number">-1</span>;
-    <span class="hljs-keyword">int</span> <span class="hljs-keyword">public</span> i256 = <span class="hljs-number">456</span>;
+    <span class="hljs-keyword">int256</span> <span class="hljs-keyword">public</span> i256 = <span class="hljs-number">456</span>;
     <span class="hljs-keyword">int</span> <span class="hljs-keyword">public</span> i = <span class="hljs-number">-123</span>; <span class="hljs-comment">// int is same as int256</span>
 
     <span class="hljs-comment">// minimum and maximum of int</span>
